% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_its.R
\name{.run_placebo_checks}
\alias{.run_placebo_checks}
\title{Internal function for placebo checks}
\usage{
.run_placebo_checks(
  y,
  X_full,
  n_pre,
  kernel_type,
  period,
  time_col,
  scale,
  interval_type,
  alpha,
  optimize,
  mixed_data,
  cat_columns,
  placebo_periods,
  verbose
)
}
\description{
Internal function for placebo checks
}
\keyword{internal}
