% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{gp_optimize}
\alias{gp_optimize}
\title{gp_optimize}
\usage{
gp_optimize(K, Y, optim.tol = 0.1)
}
\arguments{
\item{K}{kernel matrix (of covariates)}

\item{Y}{Y vector (outcome variable)}
}
\value{
\item{s2opt}{optimized s2 value}
\item{nlml}{the likelihood value at the optimal point}
}
\description{
to optimizes S2 using MLE. This function automatically runs when gp_train(..., optimize=TRUE).
}
