% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{gp_rdd}
\alias{gp_rdd}
\title{gp_rdd}
\usage{
gp_rdd(
  X,
  Y,
  cut,
  covs = NULL,
  alpha = 0.05,
  b = NULL,
  trim = FALSE,
  trim_k_value = 0.1,
  scale = TRUE
)
}
\arguments{
\item{X}{forcing variable}

\item{Y}{Y vector (outcome variable)}

\item{cut}{cut point}

\item{covs}{covariates (optional)}

\item{alpha}{confidence level (default = 0.05)}

\item{b}{bandwidth (default = NULL)}

\item{trim}{a logical value indicating whether you want to do an automatic trim at a specific value of trim_k_value (default=FALSE)}

\item{trim_k_value}{a numerical value indicating the kernel value that you want to trim above (default = 0.1)}

\item{scale}{a logical value indicating whether you want to scale the covariates (default = TRUE)}
}
\value{
\item{tau}{an estimated treatment effect}
\item{se}{the standard error of tau}
}
\description{
to perform RD using GP functions
}
\examples{
n <- 100
tau <- 3
cut <- 0
x <- rnorm(n, 0, 1)
y <- rnorm(n, 0, 1) + tau*ifelse(x>cut, 1, 0)
gp_rdd.out <- gp_rdd(x, y, cut)
gp_rdd_plot(gp_rdd.out)
}
