% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpss.R
\name{gpss}
\alias{gpss}
\title{Train a Gaussian Process Models}
\usage{
gpss(formula, data, b = NULL, s2 = 0.3, optimize = FALSE, scale = TRUE)
}
\arguments{
\item{b}{bandwidth (default = NULL)}

\item{s2}{noise or a fraction of Y not explained by X (default = 0.3)}

\item{optimize}{a logical value to indicate whether an automatic optimized value of S2 should be used. If FALSE, users must define s2. (default = FALSE)}

\item{scale}{a logical value to indicate whether covariates should be scaled. (dafault = TRUE)}
}
\value{
\item{post_mean_scaled}{posterior distribution of Y in a scaled form}
\item{post_mean_orig}{posterior distribution of Y in an original scale}
\item{post_cov_scaled}{posterior covariance matrix in a scaled form}
\item{post_cov_orig}{posterior covariance matrix in an original scale}
\item{K}{a kernel matrix of X}
\item{prior_mean_scaled}{prior distribution of mean in a scaled form}
\item{X.orig}{the original matrix or data set of X}
\item{X.init}{the original matrix or data set of X with categorical variables in an expanded form}
\item{X.init.mean}{the initial mean values of X}
\item{X.init.sd}{the initial standard deviation values of X}
\item{Y.init.mean}{the initial mean value of Y}
\item{Y.init.sd}{the initial standard deviation value of Y}
\item{K}{the kernel matrix of X}
\item{Y}{scaled Y}
\item{X}{scaled X}
\item{b}{bandwidth}
\item{s2}{sigma squared}
\item{alpha}{alpha value in Rasmussen and Williams (2006) p.19}
\item{L}{L value in Rasmussen and Williams (2006) p.19}
\item{mixed_data}{a logical value indicating whether X contains a categorical/binary variable}
\item{cat_columns}{a character or a numerical vector indicating the location of categorical/binary variables in X}
\item{cat_num}{a numerical vector indicating the location of categorical/binary variables in an expanded version of X}
\item{Xcolnames}{column names of X}
}
\description{
Train a Gaussian Process Models
}
\examples{
library(gpssinternal)
data(lalonde)

# categorical variables must be encoded as factors
dat <- transform(lalonde, race_ethnicity = factor(race_ethnicity))

# train and test sets
idx <- sample(seq_len(nrow(dat)), 500)
dat_train <- dat[idx, ]
dat_test <- dat[-idx, ]

# Fit model
mod <- gpss(re78 ~ nsw + age + educ + race_ethnicity, data = dat_train)

# predictions in the test set
p <- predict(mod, dat_test)
length(p)
head(p)
}
