% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{kernel_symmetric}
\alias{kernel_symmetric}
\title{kernel_symmetric}
\usage{
kernel_symmetric(x, b, kernel_type = "gaussian", period = NULL)
}
\value{
kernel
}
\description{
kernel_symmetric
}
\keyword{internal}
