% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mixed_data_processing}
\alias{mixed_data_processing}
\title{mixed_data_processing}
\usage{
mixed_data_processing(X, cat_columns = NULL, Xtest = NULL)
}
\arguments{
\item{X}{data set or matrix}

\item{cat_columns}{a numerical or character vector that indicates categorical variables}

\item{Xtest}{(optional) if there is testing data set separate from the training data, include the testing data}
}
\value{
data frame containing expanded form of categorical variables
}
\description{
to convert categorical variables in a data set into multiple binary variables with 1 and 0.
}
\keyword{internal}
