% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{one_hot}
\alias{one_hot}
\title{one_hot}
\usage{
one_hot(data)
}
\arguments{
\item{data}{data frame containing only categorical variables}
}
\value{
data frame containing expanded form of binary variables
}
\description{
to convert categorical variables into multiple binary variables with 1 and 0.
}
\keyword{internal}
