% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gpss}
\alias{print.gpss}
\title{print method for objects of class gpss}
\usage{
\method{print}{gpss}(x)
}
\arguments{
\item{x}{gpss object}
}
\description{
print method for objects of class gpss
}
\examples{
library(gpssinternal)
data(lalonde)
# categorical variables must be encoded as factors
dat <- transform(lalonde, race_ethnicity = factor(race_ethnicity))
# train and test sets
idx <- sample(seq_len(nrow(dat)), 500)
dat_train <- dat[idx, ]
dat_test <- dat[-idx, ]
# sample of data for speed
mod <- gpss(re78 ~ nsw + age + educ + race_ethnicity, data = dat_train)
print(mod)
}
